DSL scenario 1 DONE !!!

+++ src/sim/dsl/Scale.scala
@@ -8,5 +8,5 @@ case class Scale(val amount: Double) {
 
 	def people = amount
 
-	def invitations = amount
+	def invitations = amount.toInt
 }
